moved new logic to helper class
