+++ Portal/src/Datahub.Application/Configuration/DatahubPortalConfiguration.cs
@@ -314,3 +314,4 @@ public class StorageScanNotificationSettings
     /// </summary>
     public string DataVersion { get; set; } = "1.0";
 }
+
