+++ Portal/src/Datahub.Application/Services/Notification/IGCNotifyService.cs
@@ -16,7 +16,6 @@ public interface IGCNotifyService
     Task SendWorkspaceInactiveNotification(string email, string daysSince);
     Task SendBugReportNotification(string id, string title, string body, string email = "datasolutions-solutiondedonnees@ssc-spc.gc.ca");
     Task SendInfectedFileNotification(string email, string fileName, string workspace, string date);
-    Task SendStorageScanSuccessNotificationAsync(StorageScanSuccessNotification notification, CancellationToken cancellationToken = default);
     Task SendStorageScanSuccessEmailAsync(StorageScanSuccessNotification notification, string? recipientEmail = null, CancellationToken cancellationToken = default);
     string GetTemplateMappings(DatahubPortalConfiguration portalConfiguration);
     string GetTemplateId(string templateName, string mappingsJson);
